refactor: move mob effect API to `effect` subpackage
