+++ fabric-entity-events-v1/src/main/java/net/fabricmc/fabric/api/entity/event/v1/effect/FabricMobEffectEventHandlers.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.entity.event.v1;
+package net.fabricmc.fabric.api.entity.event.v1.effect;
 
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
