+++ fabric-entity-events-v1/src/main/java/net/fabricmc/fabric/mixin/entity/event/effect/MobEffectMixin.java
@@ -20,7 +20,7 @@ import org.spongepowered.asm.mixin.Mixin;
 
 import net.minecraft.world.effect.MobEffect;
 
-import net.fabricmc.fabric.api.entity.event.v1.FabricMobEffectEventHandlers;
+import net.fabricmc.fabric.api.entity.event.v1.effect.FabricMobEffectEventHandlers;
 
 @Mixin(MobEffect.class)
 public final class MobEffectMixin implements FabricMobEffectEventHandlers {
